requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.9
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.1
reportlab==4.0.5
